refactor: flytt roleplay state transitions ut av packages/domain (#174)
Fjerner new Date() og Math.random() fra domain. withRoleplayState, assignRoleplaySkillRoll, recordRoleplayGmSkillRoll m.fl. er flyttet til packages/rules-engine/src/encounters/roleplayTransitions.ts. startScenario() i domain krever nå eksplisitt startedAt-parameter. Tester er oppdatert og verifisert.
